DDCNL-1931: Removed unused service layers and migrated Cache specific code to CacheService and migrated the same tests from TaxAccountServiceSpec
