+++ app/uk/gov/hmrc/tai/service/CacheService.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.service
+
+import com.google.inject.Inject
+import uk.gov.hmrc.domain.Nino
+import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.tai.config.MongoConfig
+import uk.gov.hmrc.tai.connectors.{CacheConnector, CacheId}
+
+class CacheService @Inject()(mongoConfig: MongoConfig, cacheConnector: CacheConnector) {
+  def invalidateTaiCacheData(nino: Nino)(implicit hc: HeaderCarrier): Unit =
+    if (mongoConfig.mongoEnabled) {
+      cacheConnector.removeById(CacheId(nino))
+    } else {
+      ()
+    }
+}
