+++ app/uk/gov/hmrc/tai/service/benefits/BenefitsService.scala
@@ -38,13 +38,11 @@ import scala.util.control.NonFatal
 
 @Singleton
 class BenefitsService @Inject()(
-  taxAccountService: TaxAccountService,
   companyCarBenefitRepository: CompanyCarBenefitRepository,
   companyCarConnector: CompanyCarConnector,
   taxComponentService: CodingComponentService,
   iFormSubmissionService: IFormSubmissionService,
-  fileUploadService: FileUploadService,
-  pdfService: PdfService,
+  cacheService: CacheService,
   auditable: Auditor)(implicit ec: ExecutionContext) {
 
   private val logger: Logger = Logger(getClass.getName)
@@ -62,7 +60,7 @@ class BenefitsService @Inject()(
     carSeqNum: Int,
     removeCarAndFuel: WithdrawCarAndFuel)(implicit hc: HeaderCarrier): Future[String] =
     companyCarConnector.withdrawCarBenefit(nino, TaxYear(), employmentSeqNum, carSeqNum, removeCarAndFuel).andThen {
-      case Success(_) => taxAccountService.invalidateTaiCacheData(nino)
+      case Success(_) => cacheService.invalidateTaiCacheData(nino)
     }
 
   def benefits(nino: Nino, taxYear: TaxYear)(implicit hc: HeaderCarrier): Future[Benefits] =
