+++ test/uk/gov/hmrc/tai/controllers/EstimatedPayCalculatorControllerSpec.scala
@@ -17,15 +17,13 @@
 package uk.gov.hmrc.tai.controllers
 
 import org.joda.time.LocalDate
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito.when
 import play.api.libs.json.{Json, Writes}
 import play.api.test.Helpers._
 import play.api.test.{FakeHeaders, FakeRequest}
+import uk.gov.hmrc.tai.calculators.EstimatedPayCalculator
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticationPredicate
 import uk.gov.hmrc.tai.model.enums.PayFreq
-import uk.gov.hmrc.tai.model.{CalculatedPay, PayDetails}
-import uk.gov.hmrc.tai.service.TaiService
+import uk.gov.hmrc.tai.model.PayDetails
 import uk.gov.hmrc.tai.util.BaseSpec
 
 class EstimatedPayCalculatorControllerSpec extends BaseSpec {
@@ -33,12 +31,7 @@ class EstimatedPayCalculatorControllerSpec extends BaseSpec {
   "Estimated pay calculator controller" must {
     "return an OK response with CalculatedPay json" when {
       "given a valid request" in {
-
-        val mockTaiService = mock[TaiService]
-        when(mockTaiService.getCalculatedEstimatedPay(any()))
-          .thenReturn(testCalculation)
-
-        val sut = createSUT(mockTaiService)
+        val sut = createSUT()
         val response = sut.calculateFullYearEstimatedPay().apply(createRequest(payDetails))
 
         status(response) mustBe OK
@@ -47,9 +40,8 @@ class EstimatedPayCalculatorControllerSpec extends BaseSpec {
     }
     "return a bad request response" when {
       "the given request has an invalid json body" in {
-        val mockTaiService = mock[TaiService]
 
-        val sut = createSUT(mockTaiService)
+        val sut = createSUT()
         val response = sut.calculateFullYearEstimatedPay().apply(createRequest("a simple String"))
 
         status(response) mustBe BAD_REQUEST
@@ -58,9 +50,8 @@ class EstimatedPayCalculatorControllerSpec extends BaseSpec {
   }
 
   private def createSUT(
-    taiService: TaiService,
     authentication: AuthenticationPredicate = loggedInAuthenticationPredicate) =
-    new EstimatedPayCalculatorController(taiService, authentication, cc)
+    new EstimatedPayCalculatorController(authentication, cc)
 
   val date = new LocalDate(2017, 4, 14)
 
@@ -73,12 +64,7 @@ class EstimatedPayCalculatorControllerSpec extends BaseSpec {
     Some(date)
   )
 
-  private val testCalculation = CalculatedPay(
-    Some(1000),
-    Some(800),
-    Some(date),
-    Some(500)
-  )
+  private val testCalculation = EstimatedPayCalculator.calculate(payDetails)
 
   private def createRequest[T](payload: T)(implicit w: Writes[T]) = FakeRequest(
     "POST",
