+++ test/uk/gov/hmrc/tai/service/benefits/BenefitsServiceSpec.scala
@@ -19,7 +19,6 @@ package uk.gov.hmrc.tai.service.benefits
 import org.joda.time.LocalDate
 import org.mockito.ArgumentMatchers.{any, eq => meq}
 import org.mockito.Mockito._
-import org.scalatest.concurrent.ScalaFutures
 import uk.gov.hmrc.http.UnprocessableEntityException
 import uk.gov.hmrc.tai.audit.Auditor
 import uk.gov.hmrc.tai.connectors.CompanyCarConnector
@@ -31,9 +30,7 @@ import uk.gov.hmrc.tai.repositories.CompanyCarBenefitRepository
 import uk.gov.hmrc.tai.service._
 import uk.gov.hmrc.tai.util.{BaseSpec, IFormConstants}
 
-import scala.concurrent.duration._
 import scala.concurrent.Future
-import scala.language.postfixOps
 
 class BenefitsServiceSpec extends BaseSpec {
   "companyCarBenefit" must {
@@ -85,7 +82,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -121,7 +117,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -158,7 +153,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -193,7 +187,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -241,7 +234,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -301,7 +293,6 @@ class BenefitsServiceSpec extends BaseSpec {
           .thenReturn(Future.successful(taxFreeAmountComponents))
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mockCompanyCarBenefitRepository,
           mock[CompanyCarConnector],
           mockCodingComponentService)
@@ -322,17 +313,18 @@ class BenefitsServiceSpec extends BaseSpec {
         val taxYear = TaxYear()
         val removeCarAndFuel = WithdrawCarAndFuel(10, carWithdrawDate, fuelWithdrawDate)
 
-        val mockTaxAccountService = mock[TaxAccountService]
-
         val mockCompanyCarConnector = mock[CompanyCarConnector]
+
         when(
           mockCompanyCarConnector.withdrawCarBenefit(nino, taxYear, employmentSeqNum, carSeqNum, removeCarAndFuel)(hc))
           .thenReturn(Future.successful(expectedResult))
 
-        val sut = createSUT(mockTaxAccountService, mock[CompanyCarBenefitRepository], mockCompanyCarConnector)
+        val cacheService = mock[CacheService]
+
+        val sut = createSUT(mock[CompanyCarBenefitRepository], mockCompanyCarConnector, cacheService = cacheService)
         sut.withdrawCompanyCarAndFuel(nino, employmentSeqNum, carSeqNum, removeCarAndFuel)(hc).futureValue mustBe expectedResult
 
-        verify(mockTaxAccountService, times(1)).invalidateTaiCacheData(meq(nino))(any())
+        verify(cacheService, times(1)).invalidateTaiCacheData(meq(nino))(any())
         verify(mockCompanyCarConnector, times(1)).withdrawCarBenefit(any(), any(), any(), any(), any())(any())
       }
     }
@@ -533,13 +525,11 @@ class BenefitsServiceSpec extends BaseSpec {
           .sendDataEvent(any(), any())(any())
 
         val sut = createSUT(
-          mock[TaxAccountService],
           mock[CompanyCarBenefitRepository],
           mock[CompanyCarConnector],
           mock[CodingComponentService],
           mockIFormSubmissionService,
-          mock[FileUploadService],
-          mock[PdfService],
+          mock[CacheService],
           mockAuditable
         )
         val result =
@@ -574,13 +564,11 @@ class BenefitsServiceSpec extends BaseSpec {
         .sendDataEvent(any(), any())(any())
 
       val sut = createSUT(
-        mock[TaxAccountService],
         mock[CompanyCarBenefitRepository],
         mock[CompanyCarConnector],
         mock[CodingComponentService],
         mockIFormSubmissionService,
-        mock[FileUploadService],
-        mock[PdfService],
+        mock[CacheService],
         mockAuditable
       )
       sut.removeCompanyBenefits(nino, employmentId, removeCompanyBenefit)(hc).futureValue mustBe "1"
@@ -633,22 +621,18 @@ class BenefitsServiceSpec extends BaseSpec {
   )
 
   private def createSUT(
-    taxAccountService: TaxAccountService = mock[TaxAccountService],
     companyCarBenefitRepository: CompanyCarBenefitRepository = mock[CompanyCarBenefitRepository],
     companyCarConnector: CompanyCarConnector = mock[CompanyCarConnector],
     codingComponentService: CodingComponentService = mock[CodingComponentService],
     iFormSubmissionService: IFormSubmissionService = mock[IFormSubmissionService],
-    fileUploadService: FileUploadService = mock[FileUploadService],
-    pdfService: PdfService = mock[PdfService],
+    cacheService: CacheService = mock[CacheService],
     auditable: Auditor = mock[Auditor]) =
     new BenefitsService(
-      taxAccountService,
       companyCarBenefitRepository,
       companyCarConnector,
       codingComponentService,
       iFormSubmissionService,
-      fileUploadService,
-      pdfService,
+      cacheService,
       auditable
     )
 }
